Lovely lovely burds

+++ 458Bird/458Bird/Convenience.swift
@@ -22,10 +22,10 @@ Random float between min and max (inclusive).
 
 Credit to https://github.com/pNre/ExSwift
 */
-func randomFloat(min: Float = 0, max: Float) -> Float {
+func randomFloat(min: CGFloat = 0, max: CGFloat) -> CGFloat {
     let diff = max - min;
-    let rand = Float(arc4random() % (UInt32(RAND_MAX) + 1))
-    return ((rand / Float(RAND_MAX)) * diff) + min;
+    let rand = CGFloat(arc4random() % (UInt32(RAND_MAX) + 1))
+    return ((rand / CGFloat(RAND_MAX)) * diff) + min;
 }
 
 /** Perform a block of code after a delay.  Neat! */
